change triangle index to fix deletion behaviour

+++ open3d_slam/include/open3d_slam/helpers.hpp
@@ -60,7 +60,7 @@ void appendToSet(std::unordered_set<T>& vec, const std::vector<T>& other){
     vec.insert(other.begin(),other.end());
 }
 
-std::vector<Eigen::Vector3d> getPointSetFromIdx(const std::vector<int>& idxs, const PointCloud& pc);
-std::vector<Eigen::Vector3d> getPointSetFromIdx(const std::unordered_set<int>& idxs, const PointCloud& pc);
+std::vector<Eigen::Vector3d> getPointSetFromIdx(const std::vector<size_t>& idxs, const PointCloud& pc);
+std::vector<Eigen::Vector3d> getPointSetFromIdx(const std::unordered_set<size_t>& idxs, const PointCloud& pc);
 
 } /* namespace o3d_slam */
